src/imagegenerationservice/grpc_stubs.py
# -*- coding: utf-8 -*-
# Simplified gRPC implementation for ImageGenerationService
"""Generated gRPC code."""

import grpc
import demo_pb2

class ImageGenerationServiceServicer:
    """Service for generating images of cart items"""
    
    def GenerateCartImage(self, request, context):
        """Generate an image of cart items with specified style and background"""
        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
        context.set_details('Method not implemented!')
        raise NotImplementedError('Method not implemented!')

    def UploadBackground(self, request, context):
        """Upload and process background image"""
        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
        context.set_details('Method not implemented!')
        raise NotImplementedError('Method not implemented!')

    def GetImageGenerationStatus(self, request, context):
        """Get the status of an image generation request"""
        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
        context.set_details('Method not implemented!')
        raise NotImplementedError('Method not implemented!')

class CartServiceStub:
    """Client for CartService"""
    
    def __init__(self, channel):
        self.channel = channel
    
    def GetCart(self, request, timeout=None, metadata=None, with_call=False, compression=None, wait_for_ready=None, call_credentials=None, protocol_options=None):
        """Get cart for a user"""
        cart = demo_pb2.Cart()
        cart.user_id = request.user_id
        cart.items = []
        return cart

class ProductCatalogServiceStub:
    """Client for ProductCatalogService"""
    
    def __init__(self, channel):
        self.channel = channel
    
    def GetProduct(self, request, timeout=None, metadata=None, with_call=False, compression=None, wait_for_ready=None, call_credentials=None, protocol_options=None):
        """Get product details"""
        product = demo_pb2.Product()
        product.id = request.id
        product.name = f"Product {request.id}"
        product.description = f"Description for product {request.id}"
        product.picture = f"/static/img/products/{request.id}.jpg"
        product.categories = ["general"]
        return product

def add_ImageGenerationServiceServicer_to_server(servicer, server):
    """Add ImageGenerationService servicer to server"""
    rpc_method_handlers = {
        'GenerateCartImage': grpc.unary_unary_rpc_method_handler(
            servicer.GenerateCartImage,
            request_deserializer=demo_pb2.GenerateCartImageRequest.FromString,
            response_serializer=demo_pb2.GenerateCartImageResponse.SerializeToString,
        ),
        'UploadBackground': grpc.unary_unary_rpc_method_handler(
            servicer.UploadBackground,
            request_deserializer=demo_pb2.UploadBackgroundRequest.FromString,
            response_serializer=demo_pb2.UploadBackgroundResponse.SerializeToString,
        ),
        'GetImageGenerationStatus': grpc.unary_unary_rpc_method_handler(
            servicer.GetImageGenerationStatus,
            request_deserializer=demo_pb2.GetStatusRequest.FromString,
            response_serializer=demo_pb2.GetStatusResponse.SerializeToString,
        ),
    }
    generic_handler = grpc.method_handlers_generic_handler(
        'hipstershop.ImageGenerationService', rpc_method_handlers)
    server.add_generic_rpc_handlers((generic_handler,))
