src/imagegenerationservice/requirements.txt
# Copyright 2024 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# Core dependencies
grpcio==1.66.0
grpcio-tools==1.66.0
protobuf==5.28.3
flask==3.1.1
pillow==11.3.0
requests==2.32.3

# Google AI dependencies
google-generativeai==0.8.3
google-cloud-storage==2.18.0
google-cloud-secret-manager==2.24.0

# Additional utilities
python-dotenv==1.0.1
uuid==1.30
